添加一些判断防止重复写入

+++ install.sh
@@ -1,12 +1,17 @@
 # 添加vscode
+if [ ! -e /etc/apt/trusted.gpg.d/microsoft.gpg ];then
 curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > microsoft.gpg
 sudo mv microsoft.gpg /etc/apt/trusted.gpg.d/microsoft.gpg
 sudo sh -c 'echo "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main" > /etc/apt/sources.list.d/vscode.list'
+fi
 
 # 添加chrome
+if [ ! -e /etc/apt/sources.list.d/google-chrome.list ];then
 sudo wget https://repo.fdzh.org/chrome/google-chrome.list -P /etc/apt/sources.list.d/
 wget -q -O - https://dl.google.com/linux/linux_signing_key.pub  | sudo apt-key add -
+fi
 
+apt update
 
 apt install google-chrome-stable code xmonad cabal-install feh taffybar hsetroot xcompmgr numlockx libgtk-3-dev indicator-cpufreq suckless-tools sakura gobject-introspection libgirepository1.0-dev libdbusmenu-gtk3-dev libgirepository1.0-dev libwebkit2gtk-4.0-dev libgtksourceview-3.0-dev -y
